requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
requests==2.31.0
folium==0.14.0
streamlit-folium==0.15.0
geopandas==0.13.2
pandas==2.0.3
plotly==5.16.1
shapely==2.0.1
pyproj==3.6.0
python-dateutil==2.8.2